Integrate coffee log view into SPA
So that we don't have to wait for the implementation of the
Coffee Logger "business logic" layer, we stub it out, and use that on
the web page, just as we do in the test.

This gets the application "working", although the data it presents is
fake, and it won't remember any data that the user submits to it.