Implement "coffee logger" business logic layer
Provide an abstraction for recording coffee recipes, that sits on top of
a simple datastore layer.

Replace the fake coffee logger with the real one, in the SPA
integration, and fake the next layer down: the datastore.

Provide a simple loading indicator, which hides the app interface until
the datastore has fetched the coffee logger data.