Show other users' updates to the shared data in near-real-time
In the SPA integration, we set up an interval that periodically fetches
the coffee logger data. (We export the "fetch" method for this purpose).

The view was already set up to redraw the history table, when the data
was fetched. However, it would redraw too much, and the interval would
replace the text fields for adding a new record, thereby interrupting
the user.

To fix this, we set up the view to only redraw the non-interactive
portion of the interface.