Overhaul coffee log view
Update fields to be specific to the methods used in the office.

Let barista be determined once upon instantiation of the view, so that
the user doesn't have to enter it.

Remove date logic from view (todo: make this part of coffee logger).

Fix tests to reflect new view API.